Refactored parsing logic into separate file
